fix: add tests for jp arithmetic funcs
