+++ pkg/engine/jmespath/arithmetic.go
@@ -1,7 +1,6 @@
 package jmespath
 
 import (
-	"errors"
 	"fmt"
 	"math"
 	"reflect"
@@ -31,42 +30,24 @@ type Scalar struct {
 	float64
 }
 
-var errTypeMismatch = errors.New("types mismatch")
-
 func ParseArithemticOperands(arguments []interface{}, operator string) (Operand, Operand, error) {
 	op := [2]Operand{nil, nil}
-	t := [2]int{0, 0}
-
 	for i := 0; i < 2; i++ {
-		tmp, err := validateArg(divide, arguments, i, reflect.Float64)
-		if err == nil {
+		if tmp, err := validateArg(divide, arguments, i, reflect.Float64); err == nil {
 			var sc Scalar
 			sc.float64 = tmp.Float()
 			op[i] = sc
-		}
-
-		tmp, err = validateArg(divide, arguments, i, reflect.String)
-		if err == nil {
-			var q Quantity
-			q.Quantity, err = resource.ParseQuantity(tmp.String())
-			if err == nil {
-				op[i] = q
-				t[i] = 1
-			} else {
-				var d Duration
-				d.Duration, err = time.ParseDuration(tmp.String())
-				if err == nil {
-					op[i] = d
-					t[i] = 2
-				}
+		} else if tmp, err = validateArg(divide, arguments, i, reflect.String); err == nil {
+			if q, err := resource.ParseQuantity(tmp.String()); err == nil {
+				op[i] = Quantity{Quantity: q}
+			} else if d, err := time.ParseDuration(tmp.String()); err == nil {
+				op[i] = Duration{Duration: d}
 			}
 		}
 	}
-
-	if op[0] == nil || op[1] == nil || t[0]|t[1] == 3 {
+	if op[0] == nil || op[1] == nil {
 		return nil, nil, formatError(genericError, operator, "invalid operands")
 	}
-
 	return op[0], op[1], nil
 }
 
@@ -83,7 +64,7 @@ func (op1 Quantity) Add(op2 interface{}) (interface{}, error) {
 		op1.Quantity.Add(v.Quantity)
 		return op1.String(), nil
 	default:
-		return nil, errTypeMismatch
+		return nil, formatError(typeMismatchError, add)
 	}
 }
 
@@ -92,7 +73,7 @@ func (op1 Duration) Add(op2 interface{}) (interface{}, error) {
 	case Duration:
 		return (op1.Duration + v.Duration).String(), nil
 	default:
-		return nil, errTypeMismatch
+		return nil, formatError(typeMismatchError, add)
 	}
 }
 
@@ -101,7 +82,7 @@ func (op1 Scalar) Add(op2 interface{}) (interface{}, error) {
 	case Scalar:
 		return op1.float64 + v.float64, nil
 	default:
-		return nil, errTypeMismatch
+		return nil, formatError(typeMismatchError, add)
 	}
 }
 
@@ -111,7 +92,7 @@ func (op1 Quantity) Subtract(op2 interface{}) (interface{}, error) {
 		op1.Quantity.Sub(v.Quantity)
 		return op1.String(), nil
 	default:
-		return nil, errTypeMismatch
+		return nil, formatError(typeMismatchError, subtract)
 	}
 }
 
@@ -120,7 +101,7 @@ func (op1 Duration) Subtract(op2 interface{}) (interface{}, error) {
 	case Duration:
 		return (op1.Duration - v.Duration).String(), nil
 	default:
-		return nil, errTypeMismatch
+		return nil, formatError(typeMismatchError, subtract)
 	}
 }
 
@@ -129,7 +110,7 @@ func (op1 Scalar) Subtract(op2 interface{}) (interface{}, error) {
 	case Scalar:
 		return op1.float64 - v.float64, nil
 	default:
-		return nil, errTypeMismatch
+		return nil, formatError(typeMismatchError, subtract)
 	}
 }
 
@@ -154,7 +135,7 @@ func (op1 Quantity) Multiply(op2 interface{}) (interface{}, error) {
 		prod.Mul(op1.Quantity.AsDec(), q.AsDec())
 		return resource.NewDecimalQuantity(prod, op1.Quantity.Format).String(), nil
 	default:
-		return nil, errTypeMismatch
+		return nil, formatError(typeMismatchError, multiply)
 	}
 }
 
@@ -164,7 +145,7 @@ func (op1 Duration) Multiply(op2 interface{}) (interface{}, error) {
 		seconds := op1.Seconds() * v.float64
 		return time.Duration(seconds * float64(time.Second)).String(), nil
 	default:
-		return nil, errTypeMismatch
+		return nil, formatError(typeMismatchError, multiply)
 	}
 }
 
@@ -176,9 +157,9 @@ func (op1 Scalar) Multiply(op2 interface{}) (interface{}, error) {
 		return v.Multiply(op1)
 	case Duration:
 		return v.Multiply(op1)
+	default:
+		return nil, formatError(typeMismatchError, multiply)
 	}
-
-	return nil, nil
 }
 
 // Quantity / Duration			-> error
@@ -215,7 +196,7 @@ func (op1 Quantity) Divide(op2 interface{}) (interface{}, error) {
 		quo.QuoRound(op1.AsDec(), q.AsDec(), scale, inf.RoundDown)
 		return resource.NewDecimalQuantity(quo, op1.Quantity.Format).String(), nil
 	default:
-		return nil, errTypeMismatch
+		return nil, formatError(typeMismatchError, divide)
 	}
 }
 
@@ -233,7 +214,7 @@ func (op1 Duration) Divide(op2 interface{}) (interface{}, error) {
 		seconds := op1.Seconds() / v.float64
 		return time.Duration(seconds * float64(time.Second)).String(), nil
 	default:
-		return nil, errTypeMismatch
+		return nil, formatError(typeMismatchError, divide)
 	}
 }
 
@@ -245,7 +226,7 @@ func (op1 Scalar) Divide(op2 interface{}) (interface{}, error) {
 		}
 		return op1.float64 / v.float64, nil
 	default:
-		return nil, errTypeMismatch
+		return nil, formatError(typeMismatchError, divide)
 	}
 }
 
@@ -276,7 +257,7 @@ func (op1 Quantity) Modulo(op2 interface{}) (interface{}, error) {
 		}
 		return resource.NewQuantity(i1%i2, op1.Quantity.Format).String(), nil
 	default:
-		return nil, errTypeMismatch
+		return nil, formatError(typeMismatchError, modulo)
 	}
 }
 
@@ -288,7 +269,7 @@ func (op1 Duration) Modulo(op2 interface{}) (interface{}, error) {
 		}
 		return (op1.Duration % v.Duration).String(), nil
 	default:
-		return nil, errTypeMismatch
+		return nil, formatError(typeMismatchError, modulo)
 	}
 }
 
@@ -308,6 +289,6 @@ func (op1 Scalar) Modulo(op2 interface{}) (interface{}, error) {
 		}
 		return float64(val1 % val2), nil
 	default:
-		return nil, errTypeMismatch
+		return nil, formatError(typeMismatchError, modulo)
 	}
 }
